addition of justified docstrings and methodogy
